backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase
python-dotenv
pydantic==2.11.7
httpx==0.24.0
aiofiles==23.2.1
PyPDF2
python-docx==1.2.0
openpyxl
pandas
python-magic==0.4.27
Pillow==10.1.0
langchain
langchain-openai
tiktoken==0.5.2
mistralai
docling==2.44.0
google-generativeai
pydantic-settings==2.10.1
